Return the data-versions of all referenced datasets in the cjson header
Idly thinking about how to provide the functionality provided by the
x-soda2-data-out-of-date header without unconditionally involving truth
on the read path, and plumbing this information out would definitely be
part of any such thing.
