+++ common-pg/src/main/scala/com/socrata/pg/analyzer2/metatypes/InputMetaTypes.scala
@@ -76,6 +76,7 @@ object InputMetaTypes {
     private val map = new scm.HashMap[DatabaseTableName, (CopyInfo, OrderedMap[UserColumnId, ColumnInfo[CT]])]
 
     override def allCopies = map.valuesIterator.map(_._1).toVector
+    override def asMap = map.toMap
 
     // returns None if the database table name is unknown to this secondary
     def apply(dtn: DatabaseTableName): Option[(CopyInfo, OrderedMap[UserColumnId, ColumnInfo[CT]])] =
