+++ project/Dependencies.scala
@@ -19,7 +19,7 @@ object Dependencies {
     val socrataCuratorUtils = "1.2.0"
     val socrataThirdPartyUtils = "5.0.0"
     val socrataHttpCuratorBroker = "3.13.4"
-    val soqlStdlib = "4.14.40"
+    val soqlStdlib = "4.14.41"
     val typesafeConfig = "1.0.0"
     val dataCoordinator = "4.2.20"
     val typesafeScalaLogging = "3.9.2"
