+++ soql-server-pg/src/main/scala/com/socrata/pg/server/analyzer2/ProcessQuery.scala
@@ -43,6 +43,7 @@ import com.socrata.metrics.rollup.events.RollupHit
 
 import com.socrata.pg.analyzer2.{CryptProviderProvider, TransformManager, CostEstimator, CryptProvidersByDatabaseNamesProvenance, RewriteSubcolumns, Namespaces, SoQLExtraContext}
 import com.socrata.pg.analyzer2.metatypes.{CopyCache, InputMetaTypes, DatabaseMetaTypes, DatabaseNamesMetaTypes, AugmentedTableName, SoQLMetaTypesExt}
+import com.socrata.pg.analyzer2.ordering._
 import com.socrata.pg.store.{PGSecondaryUniverse, SqlUtils, RollupAnalyzer, RollupId}
 import com.socrata.pg.query.QueryResult
 import com.socrata.pg.server.CJSONWriter
@@ -407,42 +408,46 @@ object ProcessQuery {
       }
     }
 
-    ETag(etag) ~> LastModified(lastModified) ~> ContentType("application/x-socrata-gzipped-cjson") ~> Stream { rawOutputStream =>
-      for {
-        gzos <- managed(new GZIPOutputStream(rawOutputStream))
-        osw <- managed(new OutputStreamWriter(gzos, StandardCharsets.UTF_8))
-        writer <- managed(new BufferedWriter(osw))
-      } {
-        writer.write("[{")
-        for(debugFields <- debugFields) {
-          writer.write("\"debug\":%s\n ,".format(JsonUtil.renderJson(debugFields, pretty = false)))
+    ETag(etag) ~>
+      LastModified(lastModified) ~>
+      ContentType("application/x-socrata-gzipped-cjson") ~>
+      Stream { rawOutputStream =>
+        for {
+          gzos <- managed(new GZIPOutputStream(rawOutputStream))
+          osw <- managed(new OutputStreamWriter(gzos, StandardCharsets.UTF_8))
+          writer <- managed(new BufferedWriter(osw))
+        } {
+          writer.write("[{")
+          writer.write("\"dataVersions\":%s\n ,".format(JsonUtil.renderJson(copyCache.asMap.mapValues(_._1.dataVersion).toSeq.sortBy(_._1), pretty = false)))
+          for(debugFields <- debugFields) {
+            writer.write("\"debug\":%s\n ,".format(JsonUtil.renderJson(debugFields, pretty = false)))
+          }
+          writer.write("\"fingerprint\":%s\n ,".format(JString(Base64.getUrlEncoder.withoutPadding.encodeToString(etag.asBytes))))
+          writer.write("\"last_modified\":%s\n ,".format(JString(CJSONWriter.dateTimeFormat.print(lastModified))))
+          writer.write("\"locale\":%s\n ,".format(JString(locale)))
+          writer.write("\"schema\":")
+
+          @AutomaticJsonEncode
+          case class Field(c: ColumnName, t: SoQLType, h: Option[JValue], s: Boolean)
+
+          val reformattedSchema = nameAnalysis.statement.schema.values.map { case Statement.SchemaEntry(cn, typ, _hint, isSynthetic) =>
+            // not using the analysis's schema's hints because it's
+            // post-rollup, so any hints inherited from parts of the
+            // query served by a rollup will have been destroyed.
+            Field(cn, typ, columnsByName.get(cn).flatMap(_.hint), columnsByName.get(cn).map(_.isSynthetic).getOrElse(false))
+          }.toArray
+
+          JsonUtil.writeJson(writer, reformattedSchema)
+          writer.write("\n }")
+
+          for(row <- rows) {
+            writer.write("\n,")
+            CompactJsonWriter.toWriter(writer, JArray(row))
+          }
+
+          writer.write("\n]\n")
         }
-        writer.write("\"fingerprint\":%s\n ,".format(JString(Base64.getUrlEncoder.withoutPadding.encodeToString(etag.asBytes))))
-        writer.write("\"last_modified\":%s\n ,".format(JString(CJSONWriter.dateTimeFormat.print(lastModified))))
-        writer.write("\"locale\":%s\n ,".format(JString(locale)))
-        writer.write("\"schema\":")
-
-        @AutomaticJsonEncode
-        case class Field(c: ColumnName, t: SoQLType, h: Option[JValue], s: Boolean)
-
-        val reformattedSchema = nameAnalysis.statement.schema.values.map { case Statement.SchemaEntry(cn, typ, _hint, isSynthetic) =>
-          // not using the analysis's schema's hints because it's
-          // post-rollup, so any hints inherited from parts of the
-          // query served by a rollup will have been destroyed.
-          Field(cn, typ, columnsByName.get(cn).flatMap(_.hint), columnsByName.get(cn).map(_.isSynthetic).getOrElse(false))
-        }.toArray
-
-        JsonUtil.writeJson(writer, reformattedSchema)
-        writer.write("\n }")
-
-        for(row <- rows) {
-          writer.write("\n,")
-          CompactJsonWriter.toWriter(writer, JArray(row))
-        }
-
-        writer.write("\n]\n")
       }
-    }
   }
 
   def runQuery(
